% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsereg3D.ncv_modified.r
\name{sparsereg3D.ncv.m}
\alias{sparsereg3D.ncv.m}
\title{Model Evaluation}
\usage{
sparsereg3D.ncv.m(sparse.reg, lambda, w = NULL, step = FALSE)
}
\arguments{
\item{sparse.reg}{output from \code{pre.sparsereg3D} function}

\item{lambda}{vector of regularization parameter values for lasso regression}

\item{step}{logical. If TRUE stepwise procedure will be used when fitting OLS model}

\item{ols}{logical. If TRUE model will be fitted with OLS insted of using lasso}

\item{seed}{random number generator}
}
\value{
List of objects including:
\itemize{
 \item \code{model}:  list of objects containing model description
 \item \code{lambda}:   Regularization parameter value for lasso models
 \item \code{coefficients}:  Model coefficients
 \item \code{std.param}:   Standardization parameters
}

 @keywords Model evaluation
}
\description{
Model Evaluation
}
