% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsereg3D.pred.r
\name{sparsereg3D.pred}
\alias{sparsereg3D.pred}
\title{Spatial prediction at multiple soil depths.}
\usage{
sparsereg3D.pred(model.info, depths, grids, chunk.size)
}
\arguments{
\item{model.info}{output from \code{sparsereg3D.sel} function}

\item{depths}{depths at which prediction is performed}

\item{grids}{SpatialPixelsDataframe containing grids of covariates}

\item{chunk.size}{number of pixels in each chunk obtained by splitting grids in order to do parlallel processing}
}
\value{
List of prediction grids

 @keywords Spatial prediction
}
\description{
Spatial prediction at multiple soil depths.
}
