% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsereg3D.sel.r
\name{sparsereg3D.sel}
\alias{sparsereg3D.sel}
\title{Model Selection}
\usage{
sparsereg3D.sel(sparse.reg, lambda = 0, ols = FALSE, step = FALSE)
}
\arguments{
\item{sparse.reg}{output from \code{pre.sparsereg3D} function}

\item{lambda}{vector of regularization parameter values for lasso regression}

\item{ols}{should model be fitted with Ordinary Least Squares (binary)}

\item{step}{should stepwise procedure be used when fitting OLS model}

\item{seed}{random number generator}
}
\value{
List of objects including:
\itemize{
 \item \code{model}:  list of objects containing model description
 \item \code{lambda}:   Regularization parameter value for lasso models
 \item \code{coefficients}:  Model coefficients
 \item \code{std.param}:   Standardization parameters
}

 @keywords Model selection
}
\description{
Model Selection
}
