% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.sparsereg3D.r
\name{pre.sparsereg3D}
\alias{pre.sparsereg3D}
\title{Prepearing data for model selection (\code{sparsereg3D.sel} function) and model evaluation (\code{sparsereg3D.ncv} function)}
\usage{
pre.sparsereg3D(base.model, profiles, cov.grids, use.hier = FALSE,
  poly.deg = 1, num.folds = 10, num.means = 3, use.interactions = TRUE,
  standardize = TRUE, coord.trend = TRUE, seed = 321, kmean.vars = NULL,
  cum.prop = 0.9)
}
\arguments{
\item{base.model}{model description (class "formula" of form "target.variable ~ covariates + depth")}

\item{profiles}{observations of target variable (class "SoilProfileCollection")}

\item{cov.grids}{covariates (class SpatialPixelsDataFrame)}

\item{use.hier}{should hierarchy constraints be enforced (binary)}

\item{poly.deg}{degree of polynomial depth function}

\item{num.folds}{number of folds in crossvalidation}

\item{num.means}{number of clusters in k-means clustering}

\item{use.interactions}{should interactions be included in the model (binary)}

\item{standardize}{should standardization be performed (binary)}
}
\value{
List of objects including:
\itemize{
 \item \code{base.model}:  model description extended by polynomial terms if poly.deg is TRUE
 \item \code{poly.deg}:   degree of polynomial depth function
 \item \code{use.interactions}:  should interactions be used in the model (binary)
 \item \code{use.hier}:   should hierarachy constraints be enforced (binary)
 \item \code{main.effect.names}:  names of main effect covariates
 \item \code{depth.int.names}:   names of interactions with depth (e.g. slope.depth)
 \item \code{all.int.names}:  names of all interactions (not only including depth)
}

 @keywords preprocessing
}
\description{
Prepearing data for model selection (\code{sparsereg3D.sel} function) and model evaluation (\code{sparsereg3D.ncv} function)
}
