% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.sparsereg3D.r
\name{pre.sparsereg3D}
\alias{pre.sparsereg3D}
\title{Prepearing data for model selection (\code{sparsereg3D.sel} function) and model evaluation (\code{sparsereg3D.ncv} function)}
\usage{
pre.sparsereg3D(response, base.model, profiles, cov.grids, use.hier = FALSE,
  poly.deg = 1, num.folds = 10, num.means = 3, use.interactions = TRUE,
  standardize = TRUE, coord.trend = TRUE, seed = 321, kmean.vars = NULL,
  cum.prop = 0.9, s = 1)
}
\arguments{
\item{response}{Only for compositional data modeling. Default is NA. "response" specifies the names of variables in compositions. For example "response = c("sand","silt","clay")"}

\item{base.model}{model description (class "formula" of form "target.variable ~ covariates + depth" or in case of compositional data modeling just "~ covariates + depth", without target variable specified in the formula.)}

\item{profiles}{observations of target variable (class "SoilProfileCollection")}

\item{cov.grids}{covariates (class SpatialPixelsDataFrame)}

\item{use.hier}{logical. If TRUE hierarchy constraints will be enforced.}

\item{poly.deg}{degree of polynomial depth function}

\item{num.folds}{number of folds in crossvalidation}

\item{num.means}{number of clusters in k-means clustering}

\item{use.interactions}{logical. If TRUE interactions will be included in the model.}

\item{standardize}{logical. If TRUE standardization will be performed.}

\item{s}{Integer. Only for compositional data modeling. "s" specifies which variable in composition will be used as target variable for stratification. For example, s = 1 in composition "sand, silt, clay"  specifies "sand".}
}
\value{
List of objects including:
\itemize{
 \item \code{base.model}:  model description extended by polynomial terms if poly.deg is TRUE
 \item \code{poly.deg}:   degree of polynomial depth function
 \item \code{use.interactions}:  logical. If interactions are included in the model.
 \item \code{use.hier}:   logical. If hierarchy constraints  are enforced.
 \item \code{main.effect.names}:  names of main effect covariates
 \item \code{depth.int.names}:   names of interactions with depth (e.g. slope.depth)
 \item \code{all.int.names}:  names of all interactions (not only including depth)
}

 @keywords preprocessing
}
\description{
Prepearing data for model selection (\code{sparsereg3D.sel} function) and model evaluation (\code{sparsereg3D.ncv} function)
}
