% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsereg3D.ncv.r
\name{sparsereg3D.ncv}
\alias{sparsereg3D.ncv}
\title{Model Evaluation with Nested Cross-Validation}
\usage{
sparsereg3D.ncv(sparse.reg, lambda, w = NULL, step = FALSE, ols = FALSE,
  all = FALSE)
}
\arguments{
\item{sparse.reg}{output from \code{pre.sparsereg3D} function}

\item{lambda}{vector of regularization parameter values for lasso regression}

\item{step}{logical. If TRUE stepwise procedure will be used when fitting OLS model}

\item{ols}{logical. If TRUE model will be fitted with OLS insted of using lasso}

\item{all}{logical. If TRUE a detailed output will be prepared.}

\item{seed}{random number generator}
}
\value{
List of objects including:
\itemize{
 \item \code{RMSE}:  Root Mean Squared Error
 \item \code{R squared}:   Coefficient of determination
 \item Accuracy (if \code{all = TRUE}). Accuracy measures computed for each step in outer cross-valudation loop within nested cross-validation.
 \item data.prediction (if \code{all = TRUE}) Data frame containing design matrix extended with predictions obtain through nested cross-validation.
 \item models (if \code{all = TRUE}) list of length \code{nfolds} containing models that corresponds to each step in outer cross-valudation loop within nested cross-validation.
}

 @keywords Model evaluation
}
\description{
Model Evaluation with Nested Cross-Validation
}
